fix: move component to component folder
